requirements.txt
altair==5.1.2
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.0
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.5
fonttools==4.44.3
gitdb==4.0.11
GitPython==3.1.40
holidays==0.36
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
patsy==0.5.3
Pillow==10.1.0
plotly==5.18.0
pmdarima==2.0.4
prophet==1.1.5
protobuf==4.25.1
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyxlsb==1.0.10
referencing==0.31.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.0
ruptures==1.1.8
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
stanio==0.3.0
statsmodels==0.14.0
streamlit==1.28.2
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
